Evaluate logrank fitness

+++ src/Statistics.cpp
@@ -9,11 +9,11 @@
  * groupCount long and sums to length.
  */
 double logRankStatistic(const double * const targets,
-                        const int * const groups,
-                        const int * const groupCounts,
+                        const unsigned int * const groups,
+                        const unsigned int * const groupCounts,
                         const unsigned int length,
                         const unsigned int groupCount) {
-  int i, j, k;
+  unsigned int i, j, k;
   double lastTime;
   bool hasFails = false, hasCens = false;
 
@@ -21,12 +21,12 @@ double logRankStatistic(const double * const targets,
   double fails[groupCount], cens[groupCount], atRisk[groupCount],
     expectedSum[groupCount*groupCount], observedSum[groupCount],
     varianceSum[groupCount*groupCount];
-  for (i = 0; i < (int) groupCount; i++) {
+  for (i = 0; i < groupCount; i++) {
     fails[i] = 0;
     cens[i] = 0;
     atRisk[i] = groupCounts[i];
     observedSum[i] = 0;
-    for (j = i + 1; j < (int) groupCount; j++) {
+    for (j = i + 1; j < groupCount; j++) {
       expectedSum[i * groupCount + j] = 0;
       varianceSum[i * groupCount + j] = 0;
     }
@@ -37,12 +37,12 @@ double logRankStatistic(const double * const targets,
   // Times are already ordered (at least we assume so)
   // Initialize lastTime to first time
   lastTime = targets[0];
-  for (i = 0; i <= (int) length; i++) {
+  for (i = 0; i <= length; i++) {
     // If a new time is encountered, remove intermediate censored from risk
-    if ((i == (int) length || lastTime != targets[2*i]) && hasCens) {
+    if ((i == length || lastTime != targets[2*i]) && hasCens) {
       // If a new time is observed, then only censored at previous
       // times have been seen. We need to update riskgroups.
-      for (j = 0; j < (int) groupCount; j++) {
+      for (j = 0; j < groupCount; j++) {
         //printf("\nattempt with %f", cens[j]);
         atRisk[j] -= cens[j];
         cens[j] = 0;
@@ -51,16 +51,16 @@ double logRankStatistic(const double * const targets,
     }
     // When we encounter a new unique time we sum up statistics for previous
     // or we reach the end
-    if ((hasFails && targets[2*i] != lastTime) || i == (int) length) {
+    if ((hasFails && targets[2*i] != lastTime) || i == length) {
       //printf("\nUnique Time: %f", lastTime);
       // All statistics for unique time i-1 done
       // Do group stuff, always comparing j to k since k to j is equivalent
-      for (j = 0; j < (int) groupCount; j++) {
+      for (j = 0; j < groupCount; j++) {
         // Sum up all failures observed
         observedSum[j] += fails[j];
 
         // Will skip this for last group, but rest must be done
-        for (k = j + 1; k < (int) groupCount; k++) {
+        for (k = j + 1; k < groupCount; k++) {
           totalRisk = atRisk[j] + atRisk[k];
           totalFail = fails[j] + fails[k];
           // If total risk = 0, then none of the sums will have more terms added
@@ -91,7 +91,7 @@ double logRankStatistic(const double * const targets,
     }
 
     // Always update statistics, but only before end
-    if (i < (int) length){
+    if (i < length){
       // Same [failure] time as last observed failure time, just add
       // to statistics But since there might be intermediate censored
       // time, we have to update lastTime also
@@ -109,7 +109,7 @@ double logRankStatistic(const double * const targets,
     }
 
     // Update lastTime here after all group related updates
-    if (i < (int) length) {
+    if (i < length) {
       lastTime = targets[2*i];
     }
   } // End of loop
@@ -118,8 +118,8 @@ double logRankStatistic(const double * const targets,
   double sum = 0;
   // n will be groupcount * (groupcount - 1) / 2
   double n = 0;
-  for (j = 0; j < (int) groupCount; j++) {
-    for (k = j + 1; k < (int) groupCount; k++) {
+  for (j = 0; j < groupCount; j++) {
+    for (k = j + 1; k < groupCount; k++) {
       n++;
       //printf("\nObservedSum = %f", observedSum[j]);
       //printf("\nExpectedSum = %f", expectedSum[j * groupCount + k]);
